Fixed some damage still being in floats, and fixed placement of damage indicators

+++ UnityProject/Assets/Interfaces/IDamagable.cs
@@ -4,5 +4,5 @@ using System.Collections;
 
 
 public interface IDamagable{
-    void takeDamage(float damage);
+    void takeDamage(int damage);
 }
